turn back to float 16 for P100 training

+++ train_and_eval/train_stsvit.py
@@ -99,7 +99,7 @@ def train_one_epoch(model, dataloader, optimizer, criterion, device, accum_steps
         y = batch['labels'].to(device)
 
         # Amp utilized
-        with torch.autocast(device_type='cuda', dtype=torch.bfloat16):
+        with torch.autocast(device_type='cuda', dtype=torch.float16):
             # clear->logit->computeLoss->backprobagation->learn->resetLIF
             logits = model(x, dates)
             loss = criterion(logits, y) # var to store loss history, cal grad to adjust weight
